:pencil2: User Interactions for adding, updating, and removing course and opportunity instances in the degree planner :hammer: Moved StudentDegreePlannerPage-related components to a new sub-directory "DegreePlannerPage" under ui/components/student
