+++ app/imports/startup/client/routes-config.ts
@@ -39,7 +39,7 @@ import FacultyVerificationPageContainer from '../../ui/pages/faculty/FacultyVeri
 import FacultyHomePageContainer from '../../ui/pages/faculty/FacultyHomePage';
 import FacultyManageOpportunitiesPageContainer from '../../ui/pages/faculty/FacultyManageOpportunitiesPage';
 import StudentHomePageContainer from '../../ui/pages/student/StudentHomePage';
-import StudentDegreePlannerPageContainer from '../../ui/pages/student/StudentDegreePlannerPage';
+import StudentDegreePlannerPage from '../../ui/pages/student/StudentDegreePlannerPage';
 import StudentHomeAboutMePageContainer from '../../ui/pages/student/StudentHomeAboutMePage';
 import StudentHomeIcePageContainer from '../../ui/pages/student/StudentHomeIcePage';
 import StudentHomeLevelsPageContainer from '../../ui/pages/student/StudentHomeLevelsPage';
@@ -805,7 +805,7 @@ export const routes = {
     {
       path: `/${URL_ROLES.STUDENT}/${USERNAME}/degree-planner`,
       exact: true,
-      component: StudentDegreePlannerPageContainer,
+      component: StudentDegreePlannerPage,
     },
     /* ############################## MENTOR SPACE ############################## */
     {
