+++ app/imports/ui/components/student/DegreePlannerPage/DepUtilityFunctions.ts
@@ -1,6 +1,6 @@
-import { Slugs } from '../../../api/slug/SlugCollection';
-import { EXPLORER_TYPE } from '../../../startup/client/route-constants';
-import * as Router from '../shared/RouterHelperFunctions';
+import { Slugs } from '../../../../api/slug/SlugCollection';
+import { EXPLORER_TYPE } from '../../../../startup/client/route-constants';
+import * as Router from '../../shared/RouterHelperFunctions';
 
 
 export const itemSlug = (item) => Slugs.findDoc(item.slugID).name;
