+++ app/imports/ui/components/student/DegreePlannerPage/FavoriteAcademicPlansWidget.tsx
@@ -5,12 +5,12 @@ import { withTracker } from 'meteor/react-meteor-data';
 import { Link, withRouter } from 'react-router-dom';
 import _ from 'lodash';
 import { AutoForm, SelectField } from 'uniforms-semantic';
-import { IAcademicPlan } from '../../../typings/radgrad';
-import * as Router from '../shared/RouterHelperFunctions';
-import { FavoriteAcademicPlans } from '../../../api/favorite/FavoriteAcademicPlanCollection';
-import { AcademicPlans } from '../../../api/degree-plan/AcademicPlanCollection';
-import AcademicPlanViewerWidgetContainer from './AcademicPlanViewerWidget';
-import { EXPLORER_TYPE } from '../../../startup/client/route-constants';
+import { IAcademicPlan } from '../../../../typings/radgrad';
+import * as Router from '../../shared/RouterHelperFunctions';
+import { FavoriteAcademicPlans } from '../../../../api/favorite/FavoriteAcademicPlanCollection';
+import { AcademicPlans } from '../../../../api/degree-plan/AcademicPlanCollection';
+import AcademicPlanViewerWidgetContainer from '../AcademicPlanViewerWidget';
+import { EXPLORER_TYPE } from '../../../../startup/client/route-constants';
 
 interface IFavoriteAcademicPlansWidgetProps {
   match: Router.IMatchProps;
