+++ app/imports/ui/components/student/DegreePlannerPage/FavoriteCoursesWidget.tsx
@@ -3,12 +3,12 @@ import { Card, Icon, Message } from 'semantic-ui-react';
 import { withTracker } from 'meteor/react-meteor-data';
 import { Link, withRouter } from 'react-router-dom';
 import _ from 'lodash';
-import { ICourse } from '../../../typings/radgrad';
-import * as Router from '../shared/RouterHelperFunctions';
-import { FavoriteCourses } from '../../../api/favorite/FavoriteCourseCollection';
-import { Courses } from '../../../api/course/CourseCollection';
+import { ICourse } from '../../../../typings/radgrad';
+import * as Router from '../../shared/RouterHelperFunctions';
+import { FavoriteCourses } from '../../../../api/favorite/FavoriteCourseCollection';
+import { Courses } from '../../../../api/course/CourseCollection';
 import FavoriteCourseCard from './FavoriteCourseCard';
-import { EXPLORER_TYPE } from '../../../startup/client/route-constants';
+import { EXPLORER_TYPE } from '../../../../startup/client/route-constants';
 
 interface IFavoriteCoursesWidgetProps {
   match: Router.IMatchProps;
