+++ app/imports/ui/components/student/DegreePlannerPage/FavoriteOpportunitiesWidget.tsx
@@ -3,12 +3,12 @@ import { Card, Icon, Message } from 'semantic-ui-react';
 import { withTracker } from 'meteor/react-meteor-data';
 import { Link, withRouter } from 'react-router-dom';
 import _ from 'lodash';
-import { IOpportunity } from '../../../typings/radgrad';
-import { FavoriteOpportunities } from '../../../api/favorite/FavoriteOpportunityCollection';
-import * as Router from '../shared/RouterHelperFunctions';
-import { Opportunities } from '../../../api/opportunity/OpportunityCollection';
+import { IOpportunity } from '../../../../typings/radgrad';
+import { FavoriteOpportunities } from '../../../../api/favorite/FavoriteOpportunityCollection';
+import * as Router from '../../shared/RouterHelperFunctions';
+import { Opportunities } from '../../../../api/opportunity/OpportunityCollection';
 import FavoriteOpportunityCard from './FavoriteOpportunityCard';
-import { EXPLORER_TYPE } from '../../../startup/client/route-constants';
+import { EXPLORER_TYPE } from '../../../../startup/client/route-constants';
 
 interface IFavoriteOpportunitiesWidgetProps {
   match: Router.IMatchProps;
